Add support for lazy-evaulated "VarEntities" and implmentation This change adds support for a type that appends and prepends onto strings but evaluates at the end, only when a coercion to a string is invoked. This allows us to do some clever things and carry metadata (eg confidentiality) around with the string itself.
Notably, in rendering the YAML, we eseentially leave a placeholder VarEntity, but
then re-run the resulting string through the YAML renderer so that we get the
sane behaviour.

We also adjust the DockerCredentials (which uses an HTTP-Basic style auth string)
to use the new types to render, as well as the Base64ing of the Secret. This allows
a Secret to properly delay rendering of the Confidential type for a password.