Rubiks Pluggable Modules (Phase 1) (#47)
In this we switch loading of kube_objs from standard python imports to
a loader-based importer. This allows us to generate symbols "import_relative"
and "import_objects", which can make sure that we can sub-class correctly,
and allow multiple directories to seem like one "module" (because they all
end up in the kube_objs symbol table in the end). We switch to only importing
objects directly from kube_objs as a result. This doesn't extract the current
openshift-magic from the output, or actually download or check any new "module"
paths, but it sets the stage. We also fix the importer to allow import of a
multi-stage object (ie. <module>.<module> or longer).