Fix loader (#55)
Avoid saving back objects we imported in pluggable_loader
When we call import_objects(), keep track of the imported objects
and don't export them as symbols (as this can overwrite symbols
overridden in this layer. This allows us to not have to set imported
symbols to "None" explicitly as well.