SecurityContext -> SecurityContext / PodSecurityContext (#62)
***** BREAKING CHANGE *****
The original SecurityContext object was used both for container and pod
securityContext options, but it turns out these contexts are different
for the different object types. Split this out to model correctly, and
also model the capabilities and seLinuxOptions. This will break your
compiles if you're calling SecurityContext in a pod context, sorry.